feat: move logout to navbar and toggle links based on login state
