+++ frontend/src/users/Profile.module.css
@@ -23,8 +23,6 @@
   text-shadow: 0 0 0.2rem #00FFF7;
 }
 
-
-
 label {
   font-family: 'Quicksand', sans-serif;
   color: rgb(254, 255, 167);
